requirements.txt
dj-database-url==0.4.1
Django==1.11.1
django-cors-headers==2.1.0
django-tinymce==2.6.0
et-xmlfile==1.0.1
gunicorn==19.6.0
jdcal==1.3
olefile==0.44
openpyxl==2.4.9
Pillow==4.3.0
pytz==2017.2
whitenoise==3.2
